+++ scripts/metrics/go.mod
@@ -2,4 +2,10 @@ module github.com/paketo-buildpacks/github-config/scripts/metrics
 
 go 1.15
 
-require github.com/aclements/go-moremath v0.0.0-20161014184102-0ff62e0875ff
+require (
+	github.com/aclements/go-moremath v0.0.0-20161014184102-0ff62e0875ff
+	github.com/onsi/gomega v1.10.3
+	github.com/paketo-buildpacks/github-config/actions/builder/update/entrypoint v0.0.0-20201029211026-57b9ef8ef448
+	github.com/paketo-buildpacks/occam v0.0.20
+	github.com/sclevine/spec v1.4.0
+)
