+++ scripts/metrics/main.go
@@ -1,59 +1,19 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"math"
-	"net/http"
 	"os"
-	"sort"
 	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/aclements/go-moremath/stats"
+	"github.com/paketo-buildpacks/github-config/scripts/metrics/internal"
 )
 
 const dateLayout string = "2006-01-02T15:04:05Z"
 
 var orgs = []string{"paketo-buildpacks", "paketo-community"}
-
-type Repository struct {
-	Name  string `json:"name"`
-	URL   string `json:"url"`
-	Owner struct {
-		Login string `json:"login"`
-	} `json:"owner"`
-}
-
-type PullRequestUser struct {
-	Login string `json:"login"`
-}
-
-type PullRequest struct {
-	Title     string          `json:"title"`
-	Number    int64           `json:"number"`
-	MergedAt  string          `json:"merged_at,omitempty"`
-	CreatedAt string          `json:"created_at"`
-	User      PullRequestUser `json:"user"`
-	Links     struct {
-		Commits struct {
-			CommitsURL string `json:"href"`
-		} `json:"commits"`
-	} `json:"_links"`
-}
-
-type Commit struct {
-	CommitData struct {
-		Message   string `json:"message"`
-		Committer struct {
-			Name string `json:"name"`
-			Date string `json:"date"`
-		} `json:"committer"`
-	} `json:"commit"`
-}
 
 func main() {
 	var mergeTimes []float64
@@ -87,13 +47,13 @@ func main() {
 	fmt.Printf("Execution took %f seconds.\n", duration.Seconds())
 }
 
-func worker(id int, input <-chan Repository) chan float64 {
+func worker(id int, input <-chan internal.Repository) chan float64 {
 	output := make(chan float64)
 
 	go func() {
 		for repo := range input {
 			time.Sleep(time.Millisecond * 200)
-			getRepoMergeTimes(repo, output)
+			internal.GetRepoMergeTimes(repo, output)
 		}
 		close(output)
 	}()
@@ -121,122 +81,15 @@ func merge(ws ...<-chan float64) chan float64 {
 	return output
 }
 
-func getOrgReposChan(orgs []string) chan Repository {
-	output := make(chan Repository)
+func getOrgReposChan(orgs []string) chan internal.Repository {
+	output := make(chan internal.Repository)
 	go func() {
 		for _, org := range orgs {
-			for _, repo := range getOrgRepos(org) {
+			for _, repo := range internal.GetOrgRepos(org) {
 				output <- repo
 			}
 		}
 		close(output)
 	}()
 	return output
-}
-
-func getOrgRepos(org string) []Repository {
-	client := &http.Client{}
-	request, _ := http.NewRequest("GET", fmt.Sprintf("https://api.github.com/orgs/%s/repos?per_page=100", org), nil)
-	request.Header.Add("Authorization", fmt.Sprintf("token %s", os.Getenv("PAKETO_GITHUB_TOKEN")))
-
-	response, err := client.Do(request)
-	if err != nil {
-		panic(err)
-	}
-	body, _ := ioutil.ReadAll(response.Body)
-
-	repos := []Repository{}
-	err = json.Unmarshal(body, &repos)
-	if err != nil {
-		panic(string(body))
-	}
-	return repos
-}
-
-func getRepoMergeTimes(repo Repository, output chan float64) {
-	pullRequests := getClosedPullRequests(repo)
-	for _, pullRequest := range pullRequests {
-		if pullRequest.MergedAt == "" {
-			continue
-		}
-		mergedAtTime, _ := time.Parse(dateLayout, pullRequest.MergedAt)
-		if mergedAtTime.Before(time.Now().Add(-time.Hour * 30 * 24)) {
-			continue
-		}
-		if strings.Contains(pullRequest.User.Login, "bot") {
-			continue
-		}
-		if strings.Contains(pullRequest.User.Login, "nebhale") {
-			continue
-		}
-		if strings.Contains(pullRequest.Title, "rfc") {
-			continue
-		}
-		mergeTime := calculateMinutesToMerge(pullRequest)
-		fmt.Printf("Pull request %s/%s #%d by %s\ntook %f minutes to merge.\n", repo.Owner.Login, repo.Name, pullRequest.Number, pullRequest.User.Login, mergeTime)
-		output <- mergeTime
-	}
-}
-
-func getClosedPullRequests(repo Repository) []PullRequest {
-	client := &http.Client{}
-	requestURL := fmt.Sprintf("https://api.github.com/repos/%s/%s/pulls?per_page=200&state=closed", repo.Owner.Login, repo.Name)
-	request, _ := http.NewRequest("GET", requestURL, nil)
-	request.Header.Add("Authorization", fmt.Sprintf("token %s", os.Getenv("PAKETO_GITHUB_TOKEN")))
-
-	response, err := client.Do(request)
-	if err != nil {
-		panic(err)
-	}
-
-	body, _ := ioutil.ReadAll(response.Body)
-	pullRequests := []PullRequest{}
-	err = json.Unmarshal(body, &pullRequests)
-
-	if err != nil {
-		panic(fmt.Sprintf("Request: %s\nResponse: %s\n", requestURL, string(body)))
-	}
-
-	return pullRequests
-}
-
-func calculateMinutesToMerge(pullRequest PullRequest) float64 {
-	if pullRequest.MergedAt == "" {
-		panic("this pull request was never merged")
-	}
-	client := &http.Client{}
-	request, _ := http.NewRequest("GET", pullRequest.Links.Commits.CommitsURL, nil)
-	request.Header.Add("Authorization", fmt.Sprintf("token %s", os.Getenv("PAKETO_GITHUB_TOKEN")))
-
-	response, err := client.Do(request)
-	if err != nil {
-		panic(err)
-	}
-
-	body, _ := ioutil.ReadAll(response.Body)
-	pullRequestCommits := []Commit{}
-
-	err = json.Unmarshal(body, &pullRequestCommits)
-	if err != nil {
-		panic(string(body))
-	}
-
-	sort.Slice(pullRequestCommits, func(i, j int) bool {
-		iTime, _ := time.Parse(dateLayout, pullRequestCommits[i].CommitData.Committer.Date)
-		jTime, _ := time.Parse(dateLayout, pullRequestCommits[j].CommitData.Committer.Date)
-		return iTime.After(jTime)
-	})
-
-	var lastCommit Commit
-	for _, commit := range pullRequestCommits {
-		if !strings.Contains(commit.CommitData.Message, "Merge branch 'main'") {
-			lastCommit = commit
-			break
-		}
-	}
-	lastCommitTime, _ := time.Parse(dateLayout, lastCommit.CommitData.Committer.Date)
-	mergedAtTime, _ := time.Parse(dateLayout, pullRequest.MergedAt)
-
-	mergeTime := math.Round(mergedAtTime.Sub(lastCommitTime).Minutes())
-	return mergeTime
 }
